feat(aws): More EventBridge resources
