+++ plugins/source/aws/resources/services/eventbridge/api_destinations.go
@@ -1,6 +1,10 @@
 package eventbridge
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
 	"github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -12,7 +16,7 @@ func ApiDestinations() *schema.Table {
 	return &schema.Table{
 		Name:        tableName,
 		Description: `https://docs.aws.amazon.com/eventbridge/latest/APIReference/API_ApiDestination.html`,
-		Resolver:    fetchEventbridgeApiDestinations,
+		Resolver:    fetchApiDestinations,
 		Multiplex:   client.ServiceAccountRegionMultiplexer(tableName, "events"),
 		Transform:   transformers.TransformWithStruct(&types.ApiDestination{}),
 		Columns: []schema.Column{
@@ -28,4 +32,22 @@ func ApiDestinations() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchApiDestinations(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	var input eventbridge.ListApiDestinationsInput
+	c := meta.(*client.Client)
+	svc := c.Services().Eventbridge
+	for {
+		response, err := svc.ListApiDestinations(ctx, &input)
+		if err != nil {
+			return err
+		}
+		res <- response.ApiDestinations
+		if aws.ToString(response.NextToken) == "" {
+			break
+		}
+		input.NextToken = response.NextToken
+	}
+	return nil
 }
