--- plugins/source/aws/resources/services/eventbridge/connections_fetch.go
@@ -1,28 +1,0 @@
-package eventbridge
-
-import (
-	"context"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
-	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchEventbridgeConnections(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	var input eventbridge.ListConnectionsInput
-	c := meta.(*client.Client)
-	svc := c.Services().Eventbridge
-	for {
-		response, err := svc.ListConnections(ctx, &input)
-		if err != nil {
-			return err
-		}
-		res <- response.Connections
-		if aws.ToString(response.NextToken) == "" {
-			break
-		}
-		input.NextToken = response.NextToken
-	}
-	return nil
-}
