+++ plugins/source/aws/resources/services/eventbridge/endpoints.go
@@ -1,6 +1,10 @@
 package eventbridge
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
 	"github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -12,7 +16,7 @@ func Endpoints() *schema.Table {
 	return &schema.Table{
 		Name:        tableName,
 		Description: `https://docs.aws.amazon.com/eventbridge/latest/APIReference/API_Endpoint.html`,
-		Resolver:    fetchEventbridgeEndpoints,
+		Resolver:    fetchEndpoints,
 		Multiplex:   client.AccountMultiplex(tableName),
 		Transform:   transformers.TransformWithStruct(&types.Endpoint{}),
 		Columns: []schema.Column{
@@ -27,4 +31,22 @@ func Endpoints() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchEndpoints(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	var input eventbridge.ListEndpointsInput
+	c := meta.(*client.Client)
+	svc := c.Services().Eventbridge
+	for {
+		response, err := svc.ListEndpoints(ctx, &input)
+		if err != nil {
+			return err
+		}
+		res <- response.Endpoints
+		if aws.ToString(response.NextToken) == "" {
+			break
+		}
+		input.NextToken = response.NextToken
+	}
+	return nil
 }
