+++ plugins/source/aws/resources/services/eventbridge/event_bus_rules.go
@@ -1,18 +1,22 @@
 package eventbridge
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
 	"github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
 )
 
-func EventBusRules() *schema.Table {
+func eventBusRules() *schema.Table {
 	tableName := "aws_eventbridge_event_bus_rules"
 	return &schema.Table{
 		Name:        tableName,
 		Description: `https://docs.aws.amazon.com/eventbridge/latest/APIReference/API_Rule.html`,
-		Resolver:    fetchEventbridgeEventBusRules,
+		Resolver:    fetchEventBusRules,
 		Multiplex:   client.ServiceAccountRegionMultiplexer(tableName, "events"),
 		Transform:   transformers.TransformWithStruct(&types.Rule{}, transformers.WithPrimaryKeys("Arn")),
 		Columns: []schema.Column{
@@ -26,8 +30,34 @@ func EventBusRules() *schema.Table {
 			{
 				Name:     "tags",
 				Type:     schema.TypeJSON,
-				Resolver: resolveEventbridgeEventBusRuleTags,
+				Resolver: resolveEventBusRuleTags,
 			},
 		},
 	}
+}
+
+func fetchEventBusRules(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	p := parent.Item.(types.EventBus)
+	input := eventbridge.ListRulesInput{
+		EventBusName: p.Arn,
+	}
+	c := meta.(*client.Client)
+	svc := c.Services().Eventbridge
+	for {
+		response, err := svc.ListRules(ctx, &input)
+		if err != nil {
+			return err
+		}
+		res <- response.Rules
+		if aws.ToString(response.NextToken) == "" {
+			break
+		}
+		input.NextToken = response.NextToken
+	}
+	return nil
+}
+
+func resolveEventBusRuleTags(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	eventBusArn := resource.Item.(types.Rule).Arn
+	return resolveTags(ctx, meta, resource, c, *eventBusArn)
 }
