+++ plugins/source/aws/resources/services/eventbridge/replays.go
@@ -1,6 +1,11 @@
 package eventbridge
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/aws/arn"
+	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
 	"github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -12,7 +17,7 @@ func Replays() *schema.Table {
 	return &schema.Table{
 		Name:        tableName,
 		Description: `https://docs.aws.amazon.com/eventbridge/latest/APIReference/API_Replay.html`,
-		Resolver:    fetchEventbridgeReplays,
+		Resolver:    fetchReplays,
 		Multiplex:   client.ServiceAccountRegionMultiplexer(tableName, "events"),
 		Transform:   transformers.TransformWithStruct(&types.Replay{}),
 		Columns: []schema.Column{
@@ -28,4 +33,36 @@ func Replays() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchReplays(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	var input eventbridge.ListReplaysInput
+	c := meta.(*client.Client)
+	svc := c.Services().Eventbridge
+	for {
+		response, err := svc.ListReplays(ctx, &input)
+		if err != nil {
+			return err
+		}
+		res <- response.Replays
+		if aws.ToString(response.NextToken) == "" {
+			break
+		}
+		input.NextToken = response.NextToken
+	}
+	return nil
+}
+
+func resolveReplayArn(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	cl := meta.(*client.Client)
+
+	a := arn.ARN{
+		Partition: cl.Partition,
+		Service:   "events",
+		Region:    cl.Region,
+		AccountID: cl.AccountID,
+		Resource:  "replay/" + aws.ToString(resource.Item.(types.Replay).ReplayName),
+	}
+
+	return resource.Set(c.Name, a.String())
 }
